Reorganização da estrutura de pastas do projeto
