+++ app/src/main/java/com/example/telasparcial/data/AppDataBase.kt
@@ -1,9 +1,11 @@
-package com.example.telasparcial.BD
+package com.example.telasparcial.data
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.telasparcial.data.dao.ContatosDAO
+import com.example.telasparcial.data.entities.Contatos
 
 @Database(entities = [Contatos::class], version = 1)
 abstract class AppDataBase : RoomDatabase() {
