+++ app/src/main/java/com/example/telasparcial/data/entities/Contato.kt
@@ -1,4 +1,4 @@
-package com.example.telasparcial.BD
+package com.example.telasparcial.data.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
